Moved authOptions to fix bugs.
