+++ src/app/api/auth/[...nextauth]/route.ts
@@ -1,73 +1,10 @@
 import NextAuth from 'next-auth';
-import CredentialsProvider from 'next-auth/providers/credentials';
-import connectDB from '@/config/database';
-import Teacher from '@/models/Teacher';
-import bcrypt from 'bcryptjs';
-import { AuthOptions } from 'next-auth';
+import authOptions from '@/lib/auth/authOptions'; // This fixed the problem
 
 // --- THE FIX IS HERE: REMOVE 'export' ---
-const authOptions: AuthOptions = {
-	// ------------------------------------
-
-	providers: [
-		CredentialsProvider({
-			name: 'credentials',
-			credentials: {
-				email: { label: 'Email', type: 'text' },
-				password: { label: 'Password', type: 'password' },
-			},
-			async authorize(credentials) {
-				if (!credentials?.email || !credentials?.password) {
-					throw new Error('Email and password are required.');
-				}
-
-				await connectDB();
-
-				const teacher = await Teacher.findOne({
-					email: credentials.email,
-				}).select('+password');
-
-				if (!teacher || !teacher.isActive) {
-					throw new Error('Invalid credentials or inactive account.');
-				}
-
-				const isPasswordCorrect = await bcrypt.compare(
-					credentials.password,
-					teacher.password!
-				);
-
-				if (!isPasswordCorrect) {
-					throw new Error('Invalid credentials or inactive account.');
-				}
-
-				return {
-					id: teacher._id.toString(),
-					name: teacher.fullName,
-					email: teacher.email,
-				};
-			},
-		}),
-	],
-	session: {
-		strategy: 'jwt',
-	},
-	secret: process.env.AUTH_SECRET,
-	callbacks: {
-		async jwt({ token, user }) {
-			if (user) {
-				token.id = user.id;
-			}
-			return token;
-		},
-		async session({ session, token }) {
-			if (session.user) {
-				session.user.id = token.id as string;
-			}
-			return session;
-		},
-	},
-};
+// --- SECOND IMPORTANT FIX: MOVED authOptions to src/lib/auth/authOptions to allow
+// 	   both this to work, and for authOptions to be accessed in other files --- 
 
 const handler = NextAuth(authOptions);
 
-export { handler as GET, handler as POST, authOptions };
+export { handler as GET, handler as POST, };
